core: require to be in an exception gathering context for framework and plan errors
This is to avoid raising through the event processing code, which is
hard to make exception-safe (the usual symptom is that an exception would
be raised multiple times).

In addition, both #process_events and #process_events_synchronous don't
allow for recursive calls, and #clear_application_exceptions must be
called in a context where @application_exceptions was setup. This avoids
out-of-place or unbalanced framework-exception-gathering setup/cleanup.